roles/configuraciones/files/medicos-voluntarios-deploy-config/requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    compile_reqs.sh
#
argon2-cffi==19.2.0       # via -r base.in
asgiref==3.2.7            # via django
babel==2.8.0              # via django-phonenumber-field
certifi==2019.11.28       # via requests
cffi==1.14.0              # via argon2-cffi
chardet==3.0.4            # via requests
coreapi==2.3.3            # via -r base.in, drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
django-addanother==2.0.0  # via -r base.in
django-autocomplete-light==3.5.1  # via -r base.in
django-countries==6.1.2   # via -r base.in
django-dotenv==1.4.2      # via -r base.in
django-environ==0.4.5     # via -r base.in
django-extensions==2.2.9  # via -r base.in
django-filter==2.2.0      # via -r base.in
django-phonenumber-field==4.0.0  # via -r base.in
django-querysetsequence==0.12  # via -r base.in
django==3.0.5             # via -r base.in, django-addanother, django-filter, django-phonenumber-field, django-querysetsequence, djangorestframework, drf-yasg, jsonfield2
djangorestframework-gis==0.15  # via -r base.in
djangorestframework==3.11.0  # via -r base.in, djangorestframework-gis, drf-yasg
drf-yasg==1.17.1          # via -r base.in
humanfriendly==4.18       # via -r base.in
idna==2.9                 # via requests
inflection==0.3.1         # via drf-yasg
itypes==1.1.0             # via coreapi
jinja2==2.11.1            # via coreschema
jsonfield2==3.0.3         # via -r base.in
markupsafe==1.1.1         # via jinja2
packaging==20.3           # via drf-yasg
phonenumbers==8.12.0      # via -r base.in
pillow==7.1.0             # via -r base.in
psycopg2==2.8.4           # via -r base.in
pycparser==2.20           # via cffi
pyparsing==2.4.6          # via packaging
python-baseconv==1.2.2    # via -r base.in
python-dateutil==2.8.1    # via -r base.in
pytz==2019.3              # via babel, django
requests==2.23.0          # via -r base.in, coreapi
rfc3339-validator==0.1.2  # via -r base.in
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via drf-yasg
six==1.14.0               # via argon2-cffi, django-autocomplete-light, django-extensions, drf-yasg, packaging, python-dateutil, rfc3339-validator
sqlparse==0.3.1           # via django
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.8           # via requests
